create MockupImage component
