+++ client/src/components/ImageGrid.js
@@ -2,14 +2,11 @@ import React, { useCallback, useEffect, useState } from "react";
 import { useDropzone } from "react-dropzone";
 
 import Loader from "react-loader";
-import {
-  LazyLoadImage,
-  trackWindowScroll
-} from "react-lazy-load-image-component";
+import { trackWindowScroll } from "react-lazy-load-image-component";
 import ReactGA from "react-ga";
 
 import api from "../api";
-
+import MockupImage from "./MockupImage";
 import "./ImageGrid.scss";
 
 const trackUploadEvent = () => {
@@ -17,51 +14,6 @@ const trackUploadEvent = () => {
     category: "Mockup",
     action: "Upload"
   });
-};
-
-const trackDownloadEvent = template => {
-  ReactGA.event({
-    category: "Mockup",
-    action: "Download",
-    label: template
-  });
-};
-
-const MockupLoader = ({ text = "Generating preview..." }) => {
-  return (
-    <div className="loading">
-      <Loader color="white" />
-      <div className="text">{text}</div>
-    </div>
-  );
-};
-
-const MockupImage = ({ template, mockup, scrollPosition }) => {
-  const [loading, setLoading] = useState(true);
-  return (
-    <>
-      <LazyLoadImage
-        alt=""
-        src={mockup}
-        afterLoad={() => setLoading(false)}
-        threshold={500}
-        scrollPosition={scrollPosition}
-      />
-      {loading ? <MockupLoader /> : ""}
-      {mockup && !loading ? (
-        <a
-          className="button btn btn-primary"
-          href={mockup}
-          download
-          onClick={() => trackDownloadEvent(template)}
-        >
-          Download
-        </a>
-      ) : (
-        ""
-      )}
-    </>
-  );
 };
 
 const ImageGrid = ({ scrollPosition }) => {
